Perfect number
A perfect number is a number whose sum of all positive divisors except itself is equal to itself. Ask the user to enter a number. Then, code the code to find out whether the entered number is perfect or not with c #. As a screen output, print the divisors of the number and the number on the screen.